Fix a crash on core load
When adding keys to the hash table, the core checks for -1 to end the loop, but that won't always match the -1 in the enum. Both `ptr4->key` and `-1` should be the same type in order for the comparison to work.

This change allowed me to see the title screen of a game, but the game itself wouldn't load. I don't know enough about the core to know whether it's just slow, or whether there is something else going on. But it doesn't crash on launch anymore with this change.